feat: Navbar in right place & handle application theme
