+++ src/front/App.vue
@@ -5,18 +5,21 @@
       <v-container>
         <slot />
       </v-container>
+      <ThemeButton />
     </v-app>
   </div>
 </template>
 
 <script>
 
-import NavBar from '@/front/pages/NavBar.vue';
+import NavBar from '@/front/components/navbar/NavBar.vue';
+import ThemeButton from '@/front/components/themeButton/ThemeButton.vue'
 
 export default {
   name: "App",
   components: {
     NavBar,
+    ThemeButton
   },
 };
 </script>
